change double

+++ src/main/java/frc/robot/subClass/Const.java
@@ -46,7 +46,7 @@ public class Const {
     public static final double FastDrive = 0.8;
     public static final double MidDrive = 0.5;
     public static final double SlowDrive = 0.25;
-    public static final int StopDrive = 0;
+    public static final double StopDrive = 0;
     //ConveyorSpeed
     //ボールの発射(Shoot)
     public static final double BeltShoot = 0.3;
